logDev: add support for native logs -on dev builds

+++ logDev.sh
@@ -1,8 +1,11 @@
 #/bin/bash
 
+outDir=/media/data/logs/dev_logs
+
 hac_ip=10.0.0.2
 user=admin
 pass=private
+passSnmpV3=privateprivate
 
 case "$1" in
   -h|--h|-help|--help|/?)
@@ -19,11 +22,11 @@ then
 fi
 
 #get the device type/family: MSP30 / BRS50
-hac_type=`snmpget -v 3 -l authPriv -u admin -a MD5 -A privateprivate -x DES -X privateprivate ${hac_ip} hm2DevMgmtProductDescr.0 | grep -oh  "STRING: [a-zA-Z0-9]*" | sed 's/STRING: //'`
+hac_type=`snmpget -v 3 -l authPriv -u ${user} -a MD5 -A ${passSnmpV3} -x DES -X ${passSnmpV3} ${hac_ip} hm2DevMgmtProductDescr.0 | grep -oh  "STRING: [a-zA-Z0-9]*" | sed 's/STRING: //'`
 
 
 protocol=http
-hac_https=`snmpget -v 3 -l authPriv -u admin -a MD5 -A privateprivate -x DES -X privateprivate ${hac_ip} -O vq hm2WebHttpsAdminStatus.0`
+hac_https=`snmpget -v 3 -l authPriv -u ${user} -a MD5 -A ${passSnmpV3} -x DES -X ${passSnmpV3} ${hac_ip} -O vq hm2WebHttpsAdminStatus.0`
 if [ "${hac_https}" == "enable" ]
 then
   protocol=https
@@ -31,7 +34,6 @@ fi
 
 tmpFile=temp_${hac_ip}.html
 outFile=develLog_${hac_type}_${hac_ip}-$(date +"%Y.%m.%d-%H.%M.%S").html
-outDir=/media/data/logs/dev_logs
 
 
 #base64auth=`echo ${user}:${pass} | base64`
@@ -44,15 +46,24 @@ wget --http-user=${user} --http-password=${pass} --no-check-certificate  ${proto
 # check if file exist and it is not empty
 if [[ -f ${outDir}/${tmpFile} && -s ${outDir}/${tmpFile} ]]
 then
-    #echo "${tmpFile} has some data."
+  #echo "${tmpFile} has some data."
+
+  # check content: development img are not scrambled
+  logType=$(file -b --mime-type ${outDir}/${tmpFile})
+  echo "${tmpFile} is ${logType}"
+
+  if [[ ${logType} == "text/html" ]]
+  then
+    mv ${outDir}/${tmpFile}  ${outDir}/${outFile}
+  else
     unscramble ${outDir}/${tmpFile}  ${outDir}/${outFile}
-
     rm ${outDir}/${tmpFile}
+  fi
 
-    firefox ${outDir}/${outFile} &
+  firefox ${outDir}/${outFile} &
 else
-    echo "file is empty."
-    rm ${outDir}/${tmpFile}
+  echo "file is empty."
+  rm ${outDir}/${tmpFile}
 fi
